+++ tests/bootstrap.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 // require the composer autoload.php file
 require_once dirname(__DIR__) . '/vendor/autoload.php';